CCSQ Pilot/services/sharepoint_transfer/requirements.txt
Office365-REST-Python-Client
boto3